[REFACTOR] Fix duplicated lines issue detected by Sonarcloud
