+++ __mocks__/amcharts5-geodata-indonesiaLow.js
@@ -1,0 +1,4 @@
+module.exports = {
+  __esModule: true,
+  default: {}
+};
