+++ __mocks__/amcharts5-themes-Animated.js
@@ -1,0 +1,6 @@
+module.exports = {
+  __esModule: true,
+  default: {
+    new: jest.fn().mockReturnValue({})
+  }
+};
