+++ jest.config.js
@@ -11,6 +11,11 @@ module.exports = {
     // "@amcharts/amcharts5$": "<rootDir>/__mocks__/@amcharts/amcharts5.js",
     // "@amcharts/amcharts5/map$": "<rootDir>/__mocks__/@amcharts/amcharts5/map.js",
     // '^@/components/(.*)$': '<rootDir>/components/$1'
+    "\\.(css|less|scss|sass)$": "identity-obj-proxy",
+    "^@amcharts/amcharts5$": "<rootDir>/__mocks__/amcharts5.js",
+    "^@amcharts/amcharts5/map$": "<rootDir>/__mocks__/amcharts5-map.js",
+    "^@amcharts/amcharts5/themes/Animated$": "<rootDir>/__mocks__/amcharts5-themes-Animated.js",
+    "^@amcharts/amcharts5-geodata/indonesiaLow$": "<rootDir>/__mocks__/amcharts5-geodata-indonesiaLow.js"
   },
   setupFilesAfterEnv: ["<rootDir>/jest.setup.js"],
 };
