+++ setupTests.js
@@ -1,0 +1,6 @@
+import "@testing-library/jest-dom";
+
+jest.mock("@amcharts/amcharts5");
+jest.mock("@amcharts/amcharts5/map");
+jest.mock("@amcharts/amcharts5/themes/Animated");
+jest.mock("@amcharts/amcharts5-geodata/indonesiaLow");
